Fix Hi/Lo registers corruption in exception handlers
The registers are indeed save correctly (and passed around to the
relevant contexts when needed) but not restored correctly.

This causes some bugs when single steping over code (can be either C or
assembly code). It is easy to reproduce in C code but trival on assembly
codebases (steping over mul/div will trash the HI/LO registers 99% of
the time).